requirements.txt
boto3
facenet_pytorch
awslambdaric
requests
Pillow
opencv-python